IBM-Swift/Kitura#773 Moved Basic, Hash, List, Set, and Sorted Set commands to their own files. Documented base APIs
